# box/kubernetes/forklift/internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: forklift API
  version: "0.1.0"
  description: >
    Management API for forklift, a Kubernetes-native artifact repository.
    Package-format endpoints (Maven, npm, Cargo, Go, PyPI) are mounted under
    /maven, /npm, /cargo, /go and /pypi and follow each ecosystem's native protocol.
servers:
  - url: /
tags:
  - name: session
  - name: repositories
  - name: tokens
  - name: users
  - name: roles
  - name: approvals
  - name: group-mappings
paths:
  /api/v1/login:
    post:
      tags: [session]
      summary: Log in with a local username and password (sets a session cookie)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200": { description: Logged in }
        "401": { description: Invalid credentials }
  /api/v1/logout:
    post:
      tags: [session]
      summary: Clear the session cookie
      responses:
        "204": { description: Logged out }
  /api/v1/me:
    get:
      tags: [session]
      summary: Describe the current principal
      responses:
        "200": { description: Principal info }
  /api/v1/version:
    get:
      tags: [session]
      summary: Build-time version and commit metadata
      responses:
        "200": { description: Version info }
  /api/v1/repositories:
    get:
      tags: [repositories]
      summary: List repositories
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Repository" }
    post:
      tags: [repositories]
      summary: Create a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RepositoryCreate" }
      responses:
        "201": { description: Created }
        "400": { description: Validation error }
        "409": { description: Name already exists }
  /api/v1/repositories/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [repositories]
      summary: Get a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200":
          description: Repository
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Repository" }
        "404": { description: Not found }
    put:
      tags: [repositories]
      summary: Update a repository's upstream URL and config
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Updated }
    delete:
      tags: [repositories]
      summary: Delete a repository
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
  /api/v1/repositories/{id}/audit-logs:
    get:
      tags: [repositories]
      summary: List a repository's audit log (admin), newest first
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - name: event
          in: query
          schema:
            type: string
            enum: [download, upload, delete, repo.create, repo.update, repo.delete,
                   approval.request, approval.approve, approval.reject]
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  logs:
                    type: array
                    items: { $ref: "#/components/schemas/AuditLog" }
        "404": { description: Not found }
  /api/v1/tokens:
    get:
      tags: [tokens]
      summary: List the current user's personal access tokens
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Tokens }
    post:
      tags: [tokens]
      summary: Create a personal access token (plaintext returned once)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201": { description: Created }
  /api/v1/tokens/{id}:
    delete:
      tags: [tokens]
      summary: Revoke a personal access token
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Revoked }
  /api/v1/users:
    get:
      tags: [users]
      summary: List users (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Users }
    post:
      tags: [users]
      summary: Create a local user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "201": { description: Created }
  /api/v1/users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    put:
      tags: [users]
      summary: Update a user (admin) — reset password and/or toggle disabled
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password: { type: string, description: New password (local users only). }
                disabled: { type: boolean }
      responses:
        "200": { description: Updated user }
        "400": { description: Validation error (incl. disabling your own account) }
        "404": { description: Not found }
    delete:
      tags: [users]
      summary: Delete a user (admin); deleting your own account is rejected
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "204": { description: Deleted }
        "400": { description: Cannot delete your own account }
  /api/v1/users/{id}/roles:
    post:
      tags: [users]
      summary: Assign a role to a user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id: { type: integer }
      responses:
        "204": { description: Assigned }
  /api/v1/users/{id}/roles/{roleID}:
    delete:
      tags: [users]
      summary: Remove a role from a user (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: roleID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
  /api/v1/users/{id}/tokens:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      tags: [tokens]
      summary: List a user's personal access tokens (admin or auditor)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Tokens }
        "404": { description: Not found }
    post:
      tags: [tokens]
      summary: Create a personal access token for a user (admin); plaintext returned once
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TokenCreate" }
      responses:
        "201": { description: Created }
        "404": { description: Not found }
  /api/v1/users/{id}/tokens/{tokenID}:
    delete:
      tags: [tokens]
      summary: Revoke a user's personal access token (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: tokenID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Revoked }
  /api/v1/roles:
    get:
      tags: [roles]
      summary: List roles with their permissions (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Roles }
    post:
      tags: [roles]
      summary: Create a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "201": { description: Created }
  /api/v1/roles/{id}:
    delete:
      tags: [roles]
      summary: Delete a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
  /api/v1/roles/{id}/permissions:
    post:
      tags: [roles]
      summary: Add a permission to a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "201": { description: Created }
  /api/v1/roles/{id}/permissions/{permID}:
    delete:
      tags: [roles]
      summary: Remove a permission from a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: permID, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
  /api/v1/approvals:
    get:
      tags: [approvals]
      summary: List package approval requests and decisions (admin or approve permission), newest first
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: repo, in: query, schema: { type: string } }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Approval rows
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  approvals:
                    type: array
                    items: { $ref: "#/components/schemas/PackageApproval" }
    post:
      tags: [approvals]
      summary: Manually approve or reject a package (admin or approve permission on the repository), even before any request
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo, package, status]
              properties:
                repo: { type: string }
                package: { type: string, description: Canonical per-format package name. }
                status: { type: string, enum: [approved, rejected] }
                note: { type: string }
      responses:
        "201":
          description: Decision recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "400": { description: Invalid request or non-proxy repository }
        "404": { description: Repository not found }
  /api/v1/approvals/count:
    get:
      tags: [approvals]
      summary: Count approval rows (admin or approve permission), for the pending badge
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: repo, in: query, schema: { type: string } }
        - name: status
          in: query
          schema: { type: string, enum: [pending, approved, rejected] }
      responses:
        "200":
          description: Row count
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
  /api/v1/approvals/approve-all:
    post:
      tags: [approvals]
      summary: Approve every pending package in one proxy repository (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo]
              properties:
                repo: { type: string }
                note: { type: string }
      responses:
        "200":
          description: Number of pending packages approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  approved: { type: integer }
        "400": { description: Missing repo or repo is not a proxy }
        "404": { description: Repository not found }
  /api/v1/approvals/{id}/approve:
    post:
      tags: [approvals]
      summary: Approve a package (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "404": { description: Not found }
  /api/v1/approvals/{id}/reject:
    post:
      tags: [approvals]
      summary: Reject a package (admin or approve permission on the repository); cached content stops being served immediately
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                note: { type: string }
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackageApproval" }
        "404": { description: Not found }
  /api/v1/version-denies:
    get:
      tags: [approvals]
      summary: List per-version deny entries (admin or approve permission), newest first
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: repo, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 100, maximum: 500 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: Deny entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  denies:
                    type: array
                    items: { $ref: "#/components/schemas/VersionDeny" }
    post:
      tags: [approvals]
      summary: Deny one exact package version in a proxy repository (admin or approve permission on the repository); overrides package-level approval and blocks cached copies immediately
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repo, package, version]
              properties:
                repo: { type: string }
                package: { type: string, description: Canonical per-format package name. }
                version: { type: string, description: Exact version as it appears in request paths (go modules keep the "v" prefix). }
                reason: { type: string }
      responses:
        "201":
          description: Deny recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionDeny" }
        "400": { description: Invalid request or non-proxy repository }
        "404": { description: Repository not found }
  /api/v1/version-denies/{id}:
    delete:
      tags: [approvals]
      summary: Remove a deny entry (admin or approve permission on the repository)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Removed }
        "404": { description: Not found }
  /api/v1/group-mappings:
    get:
      tags: [group-mappings]
      summary: List Keycloak group-to-role mappings (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "200": { description: Mappings }
    post:
      tags: [group-mappings]
      summary: Map a Keycloak group to a role (admin)
      security: [{ basicAuth: [] }, { bearerAuth: [] }]
      responses:
        "201": { description: Created }
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Local username/password or a personal access token as the password.
    bearerAuth:
      type: http
      scheme: bearer
      description: A personal access token (flpat_...) or an OIDC bearer token.
  schemas:
    Repository:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        format: { type: string, enum: [maven, npm, cargo, go, pypi] }
        type: { type: string, enum: [hosted, proxy, group] }
        upstream_url: { type: string }
        config: { $ref: "#/components/schemas/RepoConfig" }
        artifact_count:
          type: integer
          description: Stored artifact count (list endpoint only).
        total_size:
          type: integer
          description: Total stored bytes (list endpoint only).
    RepositoryCreate:
      type: object
      required: [name, format, type]
      properties:
        name: { type: string }
        format: { type: string, enum: [maven, npm, cargo, go, pypi] }
        type: { type: string, enum: [hosted, proxy, group] }
        upstream_url: { type: string }
        config: { $ref: "#/components/schemas/RepoConfig" }
    RepoConfig:
      type: object
      properties:
        cache:
          type: object
          properties:
            enabled: { type: boolean }
            artifact_ttl: { type: string, examples: ["0", "1h"] }
            metadata_ttl: { type: string, examples: ["15m"] }
            negative_ttl: { type: string, examples: ["5m"] }
            max_size_bytes: { type: integer }
            eviction: { type: string, enum: [lru] }
        age_policy:
          type: object
          description: Supply-chain cooldown by upstream release age.
          properties:
            enabled: { type: boolean }
            min_age: { type: string, examples: ["3d"] }
            max_age: { type: string }
            action: { type: string, enum: [block, warn] }
        approval:
          type: object
          description: >
            Package approval policy (proxy repositories only). Unapproved
            packages are blocked with 403 and queued as pending requests; the
            decision unit is the whole package, version freshness stays with
            age_policy. Mode "audit" serves normally but logs and counts what
            "enforce" would have blocked.
          properties:
            enabled: { type: boolean }
            mode: { type: string, enum: [enforce, audit], default: enforce }
            auto_approve:
              type: array
              description: path.Match glob patterns of package names that bypass approval.
              items: { type: string }
              examples: [["@company/*"]]
        group:
          type: object
          description: >
            Group repository membership (type "group" only). Members are looked
            up in order and the first hit is served; groups are read-only.
          properties:
            members:
              type: array
              items: { type: string }
              examples: [["maven-hosted", "maven-central"]]
    PackageApproval:
      type: object
      properties:
        id: { type: integer }
        repo_name: { type: string }
        package: { type: string, description: Canonical per-format package name. }
        status: { type: string, enum: [pending, approved, rejected] }
        requested_by: { type: string, description: First requester, empty for anonymous. }
        decided_by: { type: string }
        note: { type: string }
        request_count: { type: integer }
        first_requested_at: { type: string, format: date-time }
        last_requested_at: { type: string, format: date-time }
        decided_at: { type: [string, "null"], format: date-time }
    VersionDeny:
      type: object
      properties:
        id: { type: integer }
        repo_name: { type: string }
        package: { type: string, description: Canonical per-format package name. }
        version: { type: string, description: Exact denied version. }
        reason: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    AuditLog:
      type: object
      properties:
        id: { type: integer }
        event:
          type: string
          enum: [download, upload, delete, repo.create, repo.update, repo.delete,
                 approval.request, approval.approve, approval.reject,
                 deny.create, deny.delete, deny.block]
        path: { type: string }
        username: { type: string, description: Empty for anonymous access. }
        method: { type: string }
        status: { type: integer }
        client_ip: { type: string }
        user_agent: { type: string }
        created_at: { type: string, format: date-time }
    TokenCreate:
      type: object
      required: [name, description, scopes, expires_in]
      properties:
        name: { type: string }
        description: { type: string }
        expires_in:
          type: string
          description: Token lifetime as a Go duration, at most one year (8760h).
          examples: ["720h"]
        scopes:
          type: array
          minItems: 1
          items:
            type: object
            required: [repo_pattern, actions]
            properties:
              repo_pattern: { type: string, examples: ["maven-*"] }
              actions:
                type: array
                minItems: 1
                items: { type: string, enum: [read, write, delete] }